Move away from mixin approach to a more composition approach
