+++ src/blogmore/generator/__init__.py
@@ -6,12 +6,7 @@ compatibility for existing imports.
 """
 
 from blogmore.generator.site import SiteGenerator
-from blogmore.generator.utils import minified_filename, paginate_posts
 
 __all__ = [
     "SiteGenerator",
-    "minified_filename",
-    "paginate_posts",
 ]
-
-### __init__.py ends here
