+++ src/blogmore/generator/features.py
@@ -1,8 +1,4 @@
-"""Mixin providing optional-feature page generation for
-[`SiteGenerator`][blogmore.generator.site.SiteGenerator].
-
-Covers feeds, search, statistics, calendar, graph, and sitemap output.
-"""
+"""Optional-feature page generation for the site generator."""
 
 from __future__ import annotations
 
@@ -13,6 +9,9 @@ from blogmore.calendar import CalendarYear, build_calendar
 from blogmore.clean_url import make_url_clean
 from blogmore.feeds import BlogFeedGenerator
 from blogmore.generator.constants import CATEGORY_DIR, TAG_DIR
+from blogmore.generator.grouping import group_posts_by_category
+from blogmore.generator.html import write_html
+from blogmore.generator.paths import canonical_url_for_path
 from blogmore.graph import GraphData, build_graph_data
 from blogmore.parser import Page, Post, post_sort_key
 from blogmore.search import write_search_index
@@ -20,18 +19,35 @@ from blogmore.sitemap import write_sitemap
 from blogmore.stats import BlogStats, compute_blog_stats
 
 if TYPE_CHECKING:
-    from blogmore.generator._protocol import GeneratorProtocol
+    from blogmore.generator.context import ContextBuilder
+    from blogmore.renderer import TemplateRenderer
+    from blogmore.site_config import SiteConfig
 
 
-class OptionalPagesMixin:
-    """Mixin that generates optional-feature pages (feeds, search, stats, calendar, graph).
+class FeatureGenerator:
+    """Generates optional-feature pages (feeds, search, stats, calendar, graph)."""
 
-    This mixin is intended to be composed into
-    [`SiteGenerator`][blogmore.generator.site.SiteGenerator] via
-    [`PagesMixin`][blogmore.generator._pages.PagesMixin].
-    """
+    def __init__(
+        self,
+        site_config: SiteConfig,
+        renderer: TemplateRenderer,
+        context_builder: ContextBuilder,
+    ) -> None:
+        """Initialize the feature generator.
 
-    def _generate_feeds(self: GeneratorProtocol, posts: list[Post]) -> None:
+        Args:
+            site_config: The site configuration.
+            renderer: The template renderer.
+            context_builder: The context builder.
+        """
+        self.site_config = site_config
+        self.renderer = renderer
+        self.context_builder = context_builder
+
+        # Feed constants - posts per feed
+        self.POSTS_PER_FEED = 20
+
+    def generate_feeds(self, posts: list[Post]) -> None:
         """Generate RSS and Atom feeds.
 
         Args:
@@ -41,14 +57,14 @@ class OptionalPagesMixin:
             output_dir=self.site_config.output_dir,
             site_title=self.site_config.site_title,
             site_url=self.site_config.site_url,
-            max_posts=self.site_config.posts_per_feed,
+            max_posts=self.POSTS_PER_FEED,
         )
 
         # Generate main index feeds
         feed_gen.generate_index_feeds(posts)
 
         # Generate category feeds
-        posts_by_category = self._group_posts_by_category(posts)
+        posts_by_category = group_posts_by_category(posts)
         # Sort posts by date for each category
         for _category_lower, (
             _category_display,
@@ -58,7 +74,7 @@ class OptionalPagesMixin:
 
         feed_gen.generate_category_feeds(posts_by_category)
 
-    def _generate_search_index(self: GeneratorProtocol, posts: list[Post]) -> None:
+    def generate_search_index(self, posts: list[Post]) -> None:
         """Generate the search index JSON file.
 
         Args:
@@ -66,19 +82,20 @@ class OptionalPagesMixin:
         """
         write_search_index(posts, self.site_config.output_dir)
 
-    def _generate_search_page(self: GeneratorProtocol, pages: list[Page]) -> None:
+    def generate_search_page(self, pages: list[Page]) -> None:
         """Generate the search page.
 
         Args:
             pages: List of static pages (for the sidebar navigation).
         """
-        context = self._get_global_context()
+        context = self.context_builder.get_global_context()
         context["pages"] = pages
         output_path = (
             self.site_config.output_dir / self.site_config.search_path.lstrip("/")
         ).resolve()
         output_path.parent.mkdir(parents=True, exist_ok=True)
-        search_url = self._get_search_url()
+
+        search_url = self.context_builder.get_search_url()
         if self.site_config.clean_urls:
             context["canonical_url"] = (
                 f"{self.site_config.site_url}{search_url}"
@@ -86,12 +103,15 @@ class OptionalPagesMixin:
                 else search_url
             )
         else:
-            context["canonical_url"] = self._canonical_url_for_path(output_path)
-        html = self.renderer.render_search_page(**context)
-        self._write_html(output_path, html)
+            context["canonical_url"] = canonical_url_for_path(
+                self.site_config, output_path
+            )
 
-    def _generate_stats_page(
-        self: GeneratorProtocol,
+        html = self.renderer.render_search_page(**context)
+        write_html(output_path, html, self.site_config.minify_html)
+
+    def generate_stats_page(
+        self,
         posts: list[Post],
         pages: list[Page],
         backlink_map: dict[str, list[Backlink]] | None = None,
@@ -102,16 +122,16 @@ class OptionalPagesMixin:
             posts: All published posts; used to compute statistics.
             pages: List of static pages (for the sidebar navigation).
             backlink_map: Optional mapping from post URL to list of
-                [`Backlink`][blogmore.backlinks.Backlink] objects.  When provided,
-                the statistics page includes a "Top Internal Links" section.
+                [`Backlink`][blogmore.backlinks.Backlink] objects.
         """
-        context = self._get_global_context()
+        context = self.context_builder.get_global_context()
         context["pages"] = pages
         output_path = (
             self.site_config.output_dir / self.site_config.stats_path.lstrip("/")
         ).resolve()
         output_path.parent.mkdir(parents=True, exist_ok=True)
-        stats_url = self._get_stats_url()
+
+        stats_url = self.context_builder.get_stats_url()
         if self.site_config.clean_urls:
             context["canonical_url"] = (
                 f"{self.site_config.site_url}{stats_url}"
@@ -119,29 +139,31 @@ class OptionalPagesMixin:
                 else stats_url
             )
         else:
-            context["canonical_url"] = self._canonical_url_for_path(output_path)
+            context["canonical_url"] = canonical_url_for_path(
+                self.site_config, output_path
+            )
+
         blog_stats: BlogStats = compute_blog_stats(
             posts, self.site_config.site_url, backlink_map
         )
         html = self.renderer.render_stats_page(stats=blog_stats, **context)
-        self._write_html(output_path, html)
+        write_html(output_path, html, self.site_config.minify_html)
 
-    def _generate_calendar_page(
-        self: GeneratorProtocol, posts: list[Post], pages: list[Page]
-    ) -> None:
+    def generate_calendar_page(self, posts: list[Post], pages: list[Page]) -> None:
         """Generate the calendar view page.
 
         Args:
             posts: All published posts; used to populate the calendar grid.
             pages: List of static pages (for the sidebar navigation).
         """
-        context = self._get_global_context()
+        context = self.context_builder.get_global_context()
         context["pages"] = pages
         output_path = (
             self.site_config.output_dir / self.site_config.calendar_path.lstrip("/")
         ).resolve()
         output_path.parent.mkdir(parents=True, exist_ok=True)
-        calendar_url = self._get_calendar_url()
+
+        calendar_url = self.context_builder.get_calendar_url()
         if self.site_config.clean_urls:
             context["canonical_url"] = (
                 f"{self.site_config.site_url}{calendar_url}"
@@ -149,38 +171,38 @@ class OptionalPagesMixin:
                 else calendar_url
             )
         else:
-            context["canonical_url"] = self._canonical_url_for_path(output_path)
-        # Determine page1_suffix for archive URL construction.  When
-        # clean_urls is enabled, strip any index filename (e.g. "index.html")
-        # so that calendar links to year/month/day archives end with a
-        # trailing slash instead of "/index.html".
+            context["canonical_url"] = canonical_url_for_path(
+                self.site_config, output_path
+            )
+
+        # Determine page1_suffix for archive URL construction.
         page1_suffix = self.site_config.page_1_path.lstrip("/")
         if self.site_config.clean_urls:
             page1_suffix = make_url_clean(f"/{page1_suffix}").lstrip("/")
+
         calendar_years: list[CalendarYear] = build_calendar(
             posts, page1_suffix, forward=self.site_config.forward_calendar
         )
         html = self.renderer.render_calendar_page(
             calendar_years=calendar_years, **context
         )
-        self._write_html(output_path, html)
+        write_html(output_path, html, self.site_config.minify_html)
 
-    def _generate_graph_page(
-        self: GeneratorProtocol, posts: list[Post], pages: list[Page]
-    ) -> None:
+    def generate_graph_page(self, posts: list[Post], pages: list[Page]) -> None:
         """Generate the post-relationship graph page.
 
         Args:
             posts: All published posts; used to build graph nodes and edges.
             pages: List of static pages (for the sidebar navigation).
         """
-        context = self._get_global_context()
+        context = self.context_builder.get_global_context()
         context["pages"] = pages
         output_path = (
             self.site_config.output_dir / self.site_config.graph_path.lstrip("/")
         ).resolve()
         output_path.parent.mkdir(parents=True, exist_ok=True)
-        graph_url = self._get_graph_url()
+
+        graph_url = self.context_builder.get_graph_url()
         if self.site_config.clean_urls:
             context["canonical_url"] = (
                 f"{self.site_config.site_url}{graph_url}"
@@ -188,7 +210,10 @@ class OptionalPagesMixin:
                 else graph_url
             )
         else:
-            context["canonical_url"] = self._canonical_url_for_path(output_path)
+            context["canonical_url"] = canonical_url_for_path(
+                self.site_config, output_path
+            )
+
         graph_data: GraphData = build_graph_data(
             posts,
             tag_dir=TAG_DIR,
@@ -198,17 +223,10 @@ class OptionalPagesMixin:
         html = self.renderer.render_graph_page(
             graph_data_json=graph_data.to_json(), **context
         )
-        self._write_html(output_path, html)
+        write_html(output_path, html, self.site_config.minify_html)
 
-    def _remove_stale_search_files(self: GeneratorProtocol) -> None:
-        """Remove search-related files left over from a previous build.
-
-        When search is disabled, any search page (at the configured
-        ``search_path`` or the default ``search.html``) and
-        ``search_index.json`` that may have been written by an earlier
-        build that had search enabled are deleted so they do not appear
-        in the output directory.
-        """
+    def remove_stale_search_files(self) -> None:
+        """Remove search-related files left over from a previous build."""
         # Always remove search_index.json (fixed location).
         stale_json = self.site_config.output_dir / "search_index.json"
         if stale_json.exists():
@@ -221,25 +239,22 @@ class OptionalPagesMixin:
         if stale_page.exists():
             stale_page.unlink()
 
-        # Also remove the default search.html location for backward
-        # compatibility (in case the user previously used the default path).
+        # Also remove the default search.html location for backward compatibility.
         default_page = (self.site_config.output_dir / "search.html").resolve()
         if default_page.exists() and default_page != stale_page:
             default_page.unlink()
 
-    def _generate_sitemap(self: GeneratorProtocol) -> None:
+    def generate_sitemap(self, extras_html_paths: frozenset[str]) -> None:
         """Generate the XML sitemap file.
 
-        Writes ``sitemap.xml`` to the root of the output directory,
-        containing an entry for every generated HTML page except the
-        configured search page and any HTML files copied verbatim from
-        the ``extras`` directory.
+        Args:
+            extras_html_paths: Relative paths of HTML files copied from extras.
         """
         write_sitemap(
             self.site_config.output_dir,
             self.site_config.site_url,
             clean_urls=self.site_config.clean_urls,
             search_path=self.site_config.search_path,
-            extra_excluded_paths=self._extras_html_paths,
+            extra_excluded_paths=extras_html_paths,
             extra_urls=self.site_config.sitemap_extras,
         )
