+++ src/blogmore/generator/pages.py
@@ -1,6 +1,4 @@
-"""Mixin providing core HTML page generation for
-[`SiteGenerator`][blogmore.generator.site.SiteGenerator].
-"""
+"""Core HTML page generation for the site generator."""
 
 from __future__ import annotations
 
@@ -10,22 +8,47 @@ from typing import TYPE_CHECKING
 from blogmore.backlinks import Backlink
 from blogmore.clean_url import make_url_clean
 from blogmore.comment_invite import build_mailto_url, get_invite_email_for_post
-from blogmore.generator._optional_pages import OptionalPagesMixin
+from blogmore.generator.html import write_html
+from blogmore.generator.paths import (
+    canonical_url_for_path,
+    get_pagination_output_path,
+    get_pagination_url,
+    pagination_prev_next,
+)
+from blogmore.generator.utils import paginate_posts
 from blogmore.parser import CUSTOM_404_HTML, Page, Post
 
 if TYPE_CHECKING:
-    from blogmore.generator._protocol import GeneratorProtocol
+    from blogmore.generator.context import ContextBuilder
+    from blogmore.renderer import TemplateRenderer
+    from blogmore.site_config import SiteConfig
 
 
-class PagesMixin(OptionalPagesMixin):
-    """Mixin that generates each type of HTML page written to the output directory.
+class PageGenerator:
+    """Generates each type of HTML page written to the output directory."""
 
-    This mixin is intended to be composed into
-    [`SiteGenerator`][blogmore.generator.site.SiteGenerator].
-    """
+    def __init__(
+        self,
+        site_config: SiteConfig,
+        renderer: TemplateRenderer,
+        context_builder: ContextBuilder,
+    ) -> None:
+        """Initialize the page generator.
 
-    def _generate_post_page(
-        self: GeneratorProtocol,
+        Args:
+            site_config: The site configuration.
+            renderer: The template renderer.
+            context_builder: The context builder.
+        """
+        self.site_config = site_config
+        self.renderer = renderer
+        self.context_builder = context_builder
+
+        # Pagination constants - posts per page for the main index
+        self.POSTS_PER_PAGE_INDEX = 10
+
+    def generate_post_page(
+        self,
         post: Post,
         all_posts: list[Post],
         pages: list[Page],
@@ -40,11 +63,9 @@ class PagesMixin(OptionalPagesMixin):
             pages: All static pages, passed to the template context.
             output_path: The pre-resolved absolute output file path for this post.
             backlinks_map: Optional mapping from post URL to list of Backlink
-                objects, built when ``with_backlinks`` is enabled.  When
-                ``None`` or when the post URL has no entry, an empty list is
-                used so the template always receives a ``backlinks`` variable.
+                objects, built when ``with_backlinks`` is enabled.
         """
-        context = self._get_global_context()
+        context = self.context_builder.get_global_context()
         context["all_posts"] = all_posts
         context["pages"] = pages
 
@@ -62,7 +83,6 @@ class PagesMixin(OptionalPagesMixin):
         )
 
         # Find previous and next posts in chronological order
-        # all_posts is already sorted by date (newest first)
         try:
             current_index = all_posts.index(post)
             # Previous post is older (higher index)
@@ -90,13 +110,14 @@ class PagesMixin(OptionalPagesMixin):
                 else post.url
             )
         else:
-            context["canonical_url"] = self._canonical_url_for_path(output_path)
-        html = self.renderer.render_post(post, **context)
-        self._write_html(output_path, html)
+            context["canonical_url"] = canonical_url_for_path(
+                self.site_config, output_path
+            )
 
-    def _generate_page(
-        self: GeneratorProtocol, page: Page, pages: list[Page], output_path: Path
-    ) -> None:
+        html = self.renderer.render_post(post, **context)
+        write_html(output_path, html, self.site_config.minify_html)
+
+    def generate_page(self, page: Page, pages: list[Page], output_path: Path) -> None:
         """Generate a single static page.
 
         Args:
@@ -104,54 +125,41 @@ class PagesMixin(OptionalPagesMixin):
             pages: All static pages, passed to the template context.
             output_path: The pre-resolved absolute output file path for this page.
         """
-        context = self._get_global_context()
+        context = self.context_builder.get_global_context()
         context["pages"] = pages
 
         output_path.parent.mkdir(parents=True, exist_ok=True)
-
-        context["canonical_url"] = self._canonical_url_for_path(output_path)
+        context["canonical_url"] = canonical_url_for_path(self.site_config, output_path)
 
         html = self.renderer.render_page(page, **context)
+        write_html(output_path, html, self.site_config.minify_html)
 
-        self._write_html(output_path, html)
-
-    def _generate_404_page(
-        self: GeneratorProtocol, page: Page, pages: list[Page]
-    ) -> None:
+    def generate_404_page(self, page: Page, pages: list[Page]) -> None:
         """Generate the custom 404 page in the root of the output directory.
 
         Args:
             page: The 404 page content to render.
             pages: All static pages, passed to the template context.
         """
-        context = self._get_global_context()
+        context = self.context_builder.get_global_context()
         context["pages"] = pages
         output_path = self.site_config.output_dir / CUSTOM_404_HTML
-        context["canonical_url"] = self._canonical_url_for_path(output_path)
+        context["canonical_url"] = canonical_url_for_path(self.site_config, output_path)
 
         html = self.renderer.render_page(page, **context)
+        write_html(output_path, html, self.site_config.minify_html)
 
-        self._write_html(output_path, html)
-
-    def _generate_index_page(
-        self: GeneratorProtocol, posts: list[Post], pages: list[Page]
-    ) -> None:
+    def generate_index_page(self, posts: list[Post], pages: list[Page]) -> None:
         """Generate the main index page with pagination.
 
         Page 1 of the main index is always written to ``index.html`` at the
-        output root, regardless of the ``page_1_path`` configuration.  This
-        guarantees that the site always has a root ``index.html``.  The
-        ``page_1_path`` setting still applies to all other paginated sections
-        (archives, tags, categories).  Pages 2 and above of the main index
-        use ``page_n_path`` as configured.
+        output root, regardless of the ``page_1_path`` configuration.
 
         Args:
             posts: All published posts, sorted newest first.
             pages: All static pages, for sidebar navigation.
         """
-        from blogmore.generator.utils import paginate_posts
-
-        context = self._get_global_context()
+        context = self.context_builder.get_global_context()
         context["pages"] = pages
 
         # Paginate posts
@@ -162,12 +170,12 @@ class PagesMixin(OptionalPagesMixin):
         total_pages = len(paginated_posts)
 
         # Page 1 of the main index is always /index.html (with clean_urls: /)
-        # regardless of page_1_path, so that the site root is never displaced.
         page1_url: str = "/index.html"
         if self.site_config.clean_urls:
             page1_url = make_url_clean(page1_url)
+
         page_urls = [page1_url] + [
-            self._get_pagination_url("", page_num)
+            get_pagination_url(self.site_config, "", page_num)
             for page_num in range(2, total_pages + 1)
         ]
 
@@ -177,11 +185,14 @@ class PagesMixin(OptionalPagesMixin):
                 output_path = self.site_config.output_dir / "index.html"
                 output_path.parent.mkdir(parents=True, exist_ok=True)
             else:
-                output_path = self._get_pagination_output_path(
-                    self.site_config.output_dir, page_num
+                output_path = get_pagination_output_path(
+                    self.site_config, self.site_config.output_dir, page_num
                 )
-            context["canonical_url"] = self._canonical_url_for_path(output_path)
-            prev_url, next_url = self._pagination_prev_next(page_num, page_urls)
+
+            context["canonical_url"] = canonical_url_for_path(
+                self.site_config, output_path
+            )
+            prev_url, next_url = pagination_prev_next(page_num, page_urls)
             context["prev_page_url"] = prev_url
             context["next_page_url"] = next_url
             context["pagination_page_urls"] = page_urls
@@ -189,24 +200,23 @@ class PagesMixin(OptionalPagesMixin):
                 page_posts, page=page_num, total_pages=total_pages, **context
             )
 
-            self._write_html(output_path, html)
+            write_html(output_path, html, self.site_config.minify_html)
 
-    def _generate_archive_page(
-        self: GeneratorProtocol, posts: list[Post], pages: list[Page]
-    ) -> None:
+    def generate_archive_page(self, posts: list[Post], pages: list[Page]) -> None:
         """Generate the archive page.
 
         Args:
             posts: All published posts.
             pages: All static pages, for sidebar navigation.
         """
-        context = self._get_global_context()
+        context = self.context_builder.get_global_context()
         context["pages"] = pages
         output_path = (
             self.site_config.output_dir / self.site_config.archive_path.lstrip("/")
         ).resolve()
         output_path.parent.mkdir(parents=True, exist_ok=True)
-        archive_url = self._get_archive_url()
+
+        archive_url = self.context_builder.get_archive_url()
         if self.site_config.clean_urls:
             context["canonical_url"] = (
                 f"{self.site_config.site_url}{archive_url}"
@@ -214,10 +224,11 @@ class PagesMixin(OptionalPagesMixin):
                 else archive_url
             )
         else:
-            context["canonical_url"] = self._canonical_url_for_path(output_path)
+            context["canonical_url"] = canonical_url_for_path(
+                self.site_config, output_path
+            )
+
         html = self.renderer.render_archive(
             posts, page=1, total_pages=1, base_path="/archive", **context
         )
-        self._write_html(output_path, html)
-
-    ### _pages.py ends here
+        write_html(output_path, html, self.site_config.minify_html)
